state: Don't recurse infinitely into subdirectories
This fixes the issue where potentially the entire /usr/lib tree is added to the state map, adding hundreds of thousands of entries. This attempts to only recurse one subdirectory to try to match what the C usysconf did.

With this commit, a first run only adds roughly 6200 entries. Still far higher than previous usysconf's 400, but at least it doesn't choke on a second run.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>